Remove test files for utils and vicinity modules
- Deleted `tests/test_utils.py` containing tests for normalization utility functions
- Removed `tests/test_vicinity.py` with comprehensive test cases for the Vicinity class
- These test files are no longer needed, likely due to refactoring or migration of tests
