+++ tests/unit/test_vicinity.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-import io
-import sys
 from pathlib import Path
 
 import numpy as np
@@ -10,7 +8,6 @@ from orjson import JSONEncodeError
 
 from vicinity import Vicinity
 from vicinity.datatypes import Backend
-from vicinity.integrations.huggingface import _MODEL_NAME_OR_PATH_PRINT_STATEMENT
 
 BackendType = tuple[Backend, str]
 
@@ -336,25 +333,3 @@ def test_vicinity_evaluate(vicinity_instance: Vicinity, vectors: np.ndarray) -> 
     vicinity_instance.backend.arguments.metric = "manhattan"
     with pytest.raises(ValueError):
         vicinity_instance.evaluate(vectors, query_vectors)
-
-
-def test_load_from_hub(vicinity_instance: Vicinity) -> None:
-    """
-    Test Vicinity.load_from_hub.
-
-    :param vicinity_instance: A Vicinity instance.
-    """
-    repo_id = "davidberenstein1957/my-vicinity-repo"
-    expected_print_statement = _MODEL_NAME_OR_PATH_PRINT_STATEMENT.split(":")[0]
-
-    # Capture the output
-    captured_output = io.StringIO()
-    sys.stdout = captured_output
-
-    Vicinity.load_from_hub(repo_id=repo_id)
-
-    # Reset redirect.
-    sys.stdout = sys.__stdout__
-
-    # Check if the expected message is in the output
-    assert expected_print_statement in captured_output.getvalue()
